chore: add latest JSON-LD context to cache
