+++ core/identity-hub-core/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
     api(project(":spi:identity-hub-spi"))
+    implementation(libs.edc.spi.jsonld)
     implementation(libs.edc.iatp.service) // JWT validator
     implementation(libs.edc.core.crypto) // JWT verifier
     implementation(libs.nimbus.jwt)
