+++ core/identity-hub-core/src/main/java/org/eclipse/edc/identityhub/core/CoreServicesExtension.java
@@ -21,6 +21,7 @@ import org.eclipse.edc.identityhub.spi.verification.AccessTokenVerifier;
 import org.eclipse.edc.identityhub.token.verification.AccessTokenVerifierImpl;
 import org.eclipse.edc.identitytrust.validation.JwtValidator;
 import org.eclipse.edc.identitytrust.verification.JwtVerifier;
+import org.eclipse.edc.jsonld.spi.JsonLd;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provider;
@@ -28,6 +29,11 @@ import org.eclipse.edc.runtime.metamodel.annotation.Setting;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.verification.jwt.SelfIssuedIdTokenVerifier;
+
+import java.net.URISyntaxException;
+
+import static org.eclipse.edc.identityhub.spi.model.IdentityHubConstants.IATP_CONTEXT_URL;
+import static org.eclipse.edc.identityhub.spi.model.IdentityHubConstants.PRESENTATION_EXCHANGE_URL;
 
 /**
  * This extension provides some core services for the IdentityHub, such as:
@@ -42,14 +48,24 @@ public class CoreServicesExtension implements ServiceExtension {
 
     @Setting(value = "Configure this IdentityHub's DID", required = true)
     public static final String OWN_DID_PROPERTY = "edc.ih.iam.id";
+    public static final String PRESENTATION_EXCHANGE_V_1_JSON = "presentation-exchange.v1.json";
+    public static final String PRESENTATION_QUERY_V_08_JSON = "presentation-query.v08.json";
     private JwtVerifier jwtVerifier;
     private JwtValidator jwtValidator;
-
     @Inject
     private DidResolverRegistry didResolverRegistry;
 
     @Inject
     private PublicKeyWrapper identityHubPublicKey;
+
+    @Inject
+    private JsonLd jsonLd;
+
+    @Override
+    public void initialize(ServiceExtensionContext context) {
+        // Setup API
+        cacheContextDocuments(getClass().getClassLoader());
+    }
 
     @Provider
     public AccessTokenVerifier createAccessTokenVerifier(ServiceExtensionContext context) {
@@ -74,5 +90,14 @@ public class CoreServicesExtension implements ServiceExtension {
 
     private String getOwnDid(ServiceExtensionContext context) {
         return context.getConfig().getString(OWN_DID_PROPERTY);
+    }
+
+    private void cacheContextDocuments(ClassLoader classLoader) {
+        try {
+            jsonLd.registerCachedDocument(PRESENTATION_EXCHANGE_URL, classLoader.getResource(PRESENTATION_EXCHANGE_V_1_JSON).toURI());
+            jsonLd.registerCachedDocument(IATP_CONTEXT_URL, classLoader.getResource(PRESENTATION_QUERY_V_08_JSON).toURI());
+        } catch (URISyntaxException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
