+++ core/identity-hub-transform/src/test/resources/presentation_query.json
@@ -4,8 +4,8 @@
       "@id": "https://w3id.org/tractusx-trust/v0.8/scope",
       "@container": "@set"
     },
-    "presentation_definition":{
-      "@id": "https://w3id.org/tractusx-trust/v0.8/presentation_definition",
+    "presentationDefinition": {
+      "@id": "https://w3id.org/tractusx-trust/v0.8/presentationDefinition",
       "@type": "@json"
     }
   }
